Update `CoreDataManager` to support initialization with a `StorageType`.
Leveraging this, the `persistentContainer` computation logic sets an `inMemoryStoreDescription` for `container.persistentStoreDescriptions`, if applicable.
This is great for unit testing and Xcode previews, because every time we create a new in-memory instance of the CoreDataManager, we start with a fresh database, and there's no overhead of having to actually persist anything in SQLLite.